feat(FRONT): adding Leaderboard , routing &.
