+++ Client/src/packages/feat-Account/Leaderboard.tsx
@@ -1,0 +1,77 @@
+import { useNavigate } from "react-router-dom"
+import { Avatar, Button } from "@mui/material"
+import { MoreVert, Chat, PersonAdd, SportsCricket } from '@mui/icons-material'
+import { Pic, players } from "./data"
+import {
+	Players,
+	Player,
+	NoPlayer,
+	Rotate,
+	/***global**/
+	Title,
+	H5,
+	Root,
+} from "./styles"
+
+function sendGameRequist(userName:string) { }
+function sendFriendRequist(userName:string) { }
+
+const Leaderboard = () => {
+	const navigate = useNavigate()
+
+	return (<Root>
+		<Title> Leaderboard </Title>
+		<Players>{
+			players.length ?
+				players.map((item, index) => (
+					<Player>
+						<div>
+							<h5> #{index + 1}</h5>
+							<p>Rank</p>
+						</div>
+						<Avatar
+							style={{ width: "60px", height: "60px" }}
+							alt={'player: ' + (index + 1)}
+							src={Pic}
+						/>
+						<img
+							style={{ width: "30px", height: "45px", margin: "5px" }}
+							alt={'player: ' + index}
+							src={item.achievement}
+						/>
+						<h5>{item.name}</h5>
+						<div style={{ display: "flex", gap: "20px" }}>
+							<div>
+								<h5>{item.ladder} </h5>
+								<H5> Ladder </H5>
+							</div>
+							<div>
+								<h5>{item.wins} </h5>
+								<H5> Wins </H5>
+							</div>
+							<div>
+								<h5>{item.loss}</h5>
+								<H5> Losses </H5>
+							</div>
+						</div>
+						<div>
+							<Button onClick={() => sendGameRequist(item.name)}><SportsCricket /></Button>
+							<Button onClick={() => navigate('/chat')}><Chat /></Button>
+							<Button onClick={() => sendFriendRequist(item.name)}><PersonAdd /></Button>
+						</div>
+					</Player>
+				)) :
+				<NoPlayer>
+					<h3> No Player is registred in <br /> the leaderboard
+						<br /><br /> <span>Play</span> and you'll here!</h3>
+					<Rotate>
+						<Button variant="contained" onClick={() => navigate('/game')}> P l a y </Button>
+					</Rotate>
+				</NoPlayer>}
+		</Players>
+	</Root>
+	)
+}
+
+// export default Leaderboard
+export { Leaderboard }
