+++ Client/src/packages/feat-Account/index.ts
@@ -1,2 +1,3 @@
-export * from "./Profile"
-export * from "./Settings"
+export { Profile } from "./Profile"
+export { Settings } from "./Settings"
+export { Leaderboard } from "./Leaderboard"
